fix: fix hp type int to float

+++ Assets/Scripts/3.31/SetHpOfChracters.cs
@@ -5,17 +5,17 @@ using UnityEngine.UI;
 
 public class SetHpOfChracters 
 {
-    private static int FULL_HP = 100;
+    private static float FULL_HP = 100;
 
     private static Slider playerHpBar = GameObject.Find("PlayerHpBar").GetComponent<Slider>();
     private static Slider enemyHpBar = GameObject.Find("EnemyHpBar").GetComponent<Slider>();
 
-    public static void updatePlayerHp(int playerHp)
+    public static void updatePlayerHp(float playerHp)
     {
         playerHpBar.value = playerHp / FULL_HP;
     }
 
-    public static void updateEnemyHp(int enemyHp)
+    public static void updateEnemyHp(float enemyHp)
     {
         enemyHpBar.value = enemyHp / FULL_HP;
     }
